backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.2.0
python-decouple==3.8
qrcode==7.4.2
reportlab==4.1.0
PyPDF2==3.0.1
geopy==2.4.1
django-storages==1.14.2
boto3==1.34.51
drf-yasg==1.21.7
gunicorn==21.2.0
